Moved dict utilities to dict.rs.
